pyproject.toml
# This file is part of "austin-web" which is released under GPL.
#
# See file LICENCE or go to http://www.gnu.org/licenses/ for full license
# details.
#
# austin-web is a Python wrapper around Austin, the CPython frame stack
# sampler.
#
# Copyright (c) 2018-2020 Gabriele N. Tornetta <phoenix1987@gmail.com>.
# All rights reserved.
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.

[project]
name = "austin-web"
description = "Flame graph web application for Austin"
authors = [{ name = "Gabriele N. Tornetta", email = "phoenix1987@gmail.com" }]
license = "GPL-3.0-or-later"
readme = "README.md"

keywords = ["performance", "profiling", "testing", "development"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

requires-python = ">=3.9"
dependencies = ["aiohttp~=3.6", "austin-python~=1.7", "halo~=0.0.29"]

dynamic = ["version"]

[project.urls]
homepage = "https://github.com/P403n1x87/austin-web"
documentation = "https://austin-web.readthedocs.io"
repository = "https://github.com/P403n1x87/austin-web"
issues = "https://github.com/P403n1x87/austin-web/issues"

[project.scripts]
austin-web = "austin_web.__main__:main"

[tool.hatch.envs.tests]
template = "tests"
dependencies = [
  "austin-dist~=3.6",
  "pytest>=5.4.2",
  "pytest-asyncio",
  "pytest-cov>=2.8.1",
  "requests",
]
[tool.hatch.envs.tests.scripts]
tests = "pytest --cov=austin --cov-report=term-missing --cov-report=xml {args}"

[[tool.hatch.envs.tests.matrix]]
python = ["3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.checks]
python = "3.10"
template = "checks"
dependencies = [
  "mypy~=0.770",
  "flake8~=5.0.4",
  "flake8-annotations~=2.9.1",
  "flake8-black",
  "flake8-bugbear~=22.9.23",
  "flake8-docstrings~=1.6.0",
  "flake8-import-order~=0.18.1",
  "flake8-isort~=5.0.0",
]

[tool.hatch.envs.checks.scripts]
typing = "mypy --show-error-codes --install-types --non-interactive {args} austin_web/ test/"
linting = "flake8 {args} austin_web/ test/ "

[tool.hatch.envs.coverage]
python = "3.10"
template = "coverage"
dependencies = ["coverage[toml]", "codecov"]

[tool.hatch.envs.coverage.scripts]
cov = "coverage xml --fail-under=50"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.build.targets.wheel]
packages = ["austin_web"]

[tool.coverage.run]
branch = true
source = ["austin_web"]

[tool.coverage.report]
show_missing = true

[tool.isort]
force_single_line = true
lines_after_imports = 2
profile = "black"

[tool.mypy]
ignore_missing_imports = true
exclude = []

# [[tool.mypy.overrides]]
# module = []
# ignore_errors = true

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"
